+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed popup wrong position issues in some conditions (#16)
+
 ## 1.0.0-beta.20 (09-09-2019)
 
 - Moved wrapping static outside (to `@discoveryjs/cli`)